+++ components/core/ui/ktx/src/main/java/com/flipperdevices/core/ui/ktx/OrangeAppBar.kt
@@ -1,6 +1,5 @@
 package com.flipperdevices.core.ui.ktx
 
-import com.flipperdevices.core.ui.res.R as DesignSystem
 import androidx.annotation.DrawableRes
 import androidx.annotation.StringRes
 import androidx.compose.foundation.Image
@@ -20,6 +19,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
+import com.flipperdevices.core.ui.res.R as DesignSystem
 import com.flipperdevices.core.ui.theme.LocalPallet
 import com.flipperdevices.core.ui.theme.LocalTypography
 
@@ -75,7 +75,6 @@ fun OrangeAppBar(
         }
     }
 }
-
 
 @Composable
 fun OrangeAppBarWithIcon(