+++ components/faphub/category/impl/src/main/java/com/flipperdevices/faphub/category/impl/api/FapHubCategoryApiImpl.kt
@@ -44,7 +44,6 @@ class FapHubCategoryApiImpl @Inject constructor(
                 onBack = navController::popBackStack,
                 onOpenSearch = { navController.navigate(searchEntryApi.start()) },
                 onOpenFapItem = {
-
                 }
             )
         }