+++ components/hub/impl/src/main/java/com/flipperdevices/hub/impl/fragments/HubFragment.kt
@@ -24,7 +24,6 @@ class HubFragment : ComposeFragment(), OnBackPressListener {
     @Inject
     lateinit var hubFeatureEntry: HubFeatureEntry
 
-
     init {
         ComponentHolder.component<HubComponent>().inject(this)
     }